Add some bitwise and booleans ops as well as modulo op plus basic typing.
This is the first commit to introduce a real notion of types to Mu.
Before this everything was hardcoded to I32. Now muc can generate i1 and
i32 types in MLIR. Most of the other types are still not supported but
the typing table is in the parser at least.

There was a need to add a typing table because of Bison's nature of
passing everything around as a void pointer. Unfortunate but it is
possible that the notion of a mu::ast::ASTTypeContainer along with the
type table will make it possible to look custom types up more easily (ie
types created using structs or typealiases, which are still not
accounted for at all in the parser but could come some day).